Počelo pisanje kodera. Niš se još ne događa kad se pokrene. Tek pišem učitavanje yuv datoteke.

+++ fer_h264/fer_h264/h264_globals.h
@@ -201,7 +201,7 @@ typedef struct {
 extern frame_type frame;
 
 // The absolute frame count starting with 1
-extern int frameCount;
+extern unsigned long frameCount;
 
 //frame/4 dimenzije
 extern int *Intra4x4PredMode;
